+++ src/layouts/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Helmet from 'react-helmet';
+import Header from './Header';
 import './index.scss'
 import '../components/css/custom.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Header from './Header';
 import '../assets/images/512.png';
 
 const closePopup = () => {
