+++ src/pages/index.js
@@ -1,11 +1,10 @@
 import React, { Component } from "react";
 import _ from 'lodash'
-import ProductList from "../components/ProductList";
 import Img from 'gatsby-image';
 import CorouselItem from "../components/CorouselItem";
 import HomeStep from "../components/HomeStep";
 import './home.scss';
-import ProductCard from "../components/ProductCard";
+import ProductCard from "../components/ProductList/ProductCard.js";
 
 const planSteps = [
   {
